refactor: 컴포넌트도 파일로 분리, index.html 수정
컴포넌트도 파일로 분리
- 렌더링을 위한 컴포넌트(base component 포함)를 각각의 파일로 분리함

index.html 수정
- index.html이 app.js 파일이 아니라 여러 파일을 하나의 파일로 번들링한 bundle.js을 가져오도록 함
